requirements.txt
requests>=2.31.0
pycryptodome>=3.19.0
Pillow